# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Retrieve a list of appointments",
                "description": "Fetch a paginated list of appointments with optional filters for patient name, doctor name, clinic name, status, and appointment time.",
                "operationId": "c68dab577f29c04c4422e6cb86fa36ce",
                "parameters": [
                    {
                        "name": "patient_name",
                        "in": "query",
                        "description": "Filter by patient name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "doctor_name",
                        "in": "query",
                        "description": "Filter by doctor name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "clinic_name",
                        "in": "query",
                        "description": "Filter by clinic name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by appointment status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "description": "Filter appointments starting from this time",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "description": "Filter appointments ending at this time",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "appointment_time",
                                "status"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Appointment"
                                            }
                                        },
                                        "meta": {
                                            "description": "Pagination metadata",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Schedule a new appointment for a patient with a doctor at a specific clinic.",
                "operationId": "0cfd7a3482605b745b634d591966d8bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clinic_id",
                                    "doctor_id",
                                    "appointment_time"
                                ],
                                "properties": {
                                    "clinic_id": {
                                        "description": "ID of the clinic",
                                        "type": "integer"
                                    },
                                    "doctor_id": {
                                        "description": "ID of the doctor",
                                        "type": "integer"
                                    },
                                    "appointment_time": {
                                        "description": "Scheduled time for the appointment",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment created successfully"
                                        },
                                        "appointment": {
                                            "$ref": "#/components/schemas/Appointment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Appointment": {
                "properties": {
                    "id": {
                        "description": "Appointment ID",
                        "type": "integer"
                    },
                    "patient_id": {
                        "description": "Patient ID",
                        "type": "integer"
                    },
                    "doctor_id": {
                        "description": "Doctor ID",
                        "type": "integer"
                    },
                    "clinic_id": {
                        "description": "Clinic ID",
                        "type": "integer"
                    },
                    "appointment_time": {
                        "description": "Appointment time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "description": "Appointment status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}